+++ integration-tests/oidc-code-flow/src/main/resources/application.properties
@@ -121,6 +121,7 @@ quarkus.oidc.tenant-nonce.client-id=quarkus-app
 quarkus.oidc.tenant-nonce.credentials.secret=secret
 quarkus.oidc.tenant-nonce.authentication.scopes=profile,email,phone
 quarkus.oidc.tenant-nonce.authentication.extra-params.max-age=60
+quarkus.oidc.tenant-nonce.authentication.redirect-path=/tenant-nonce
 quarkus.oidc.tenant-nonce.application-type=web-app
 quarkus.oidc.tenant-nonce.authentication.nonce-required=true
 quarkus.oidc.tenant-nonce.authentication.state-secret=eUk1p7UB3nFiXZGUXi0uph1Y9p34YhBU
