+++ integration-tests/oidc-wiremock/src/main/resources/application.properties
@@ -90,6 +90,7 @@ quarkus.oidc.bearer-user-info-github-service.credentials.secret=AyM1SysPpbyDfgZl
 quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.provider=github
 quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.auth-server-url=${keycloak.url}/realms/quarkus/
 quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.authorization-path=/
+quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.token-path=access_token_refreshed
 quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.user-info-path=protocol/openid-connect/userinfo
 quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.jwks-path=${keycloak.url}/realms/quarkus/protocol/openid-connect/certs
 quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.code-grant.extra-params.extra-param=extra-param-value
@@ -98,7 +99,8 @@ quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.cache-user-info-in-idt
 quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.token.refresh-token-time-skew=298
 quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.authentication.verify-access-token=true
 quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.client-id=quarkus-web-app
-quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.credentials.secret=AyM1SysPpbyDfgZld3umj1qzKObwVMkoqQ-EstJQLr_T-1qS0gZH75aKtMN3Yj0iPS4hcgUuTwjAzZr1Z9CAow
+quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.credentials.client-secret.value=AyM1SysPpbyDfgZld3umj1qzKObwVMkoqQ-EstJQLr_T-1qS0gZH75aKtMN3Yj0iPS4hcgUuTwjAzZr1Z9CAow
+quarkus.oidc.code-flow-user-info-github-cached-in-idtoken.credentials.client-secret.method=query
 
 
 quarkus.oidc.code-flow-token-introspection.provider=github
